backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Azure
azure-storage-blob==12.19.0
azure-monitor-opentelemetry==1.2.0
opentelemetry-instrumentation-fastapi==0.43b0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client (for IP geolocation)
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
cachetools==5.3.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
